Add files via upload
This is a shiny R project!

Information form the open-source database is integrated and joined into a single .csv file based on Information Gain of factors affecting the load forecast. 
RHHFulFinal.csv is the training dataset that is cleaned and preprocessed. 
RHHFulFinaltest.csv is a test dataset. 

The three R files together form the dashboard: 
ui.R - generates interactive UI 
server.R - The backend of the application responsible for integrating the database and UI.
app.R -  launches the dashboard app